+++ BambooTracker/bamboo_tracker.hpp
@@ -324,6 +324,19 @@ public:
 	void addSong(SongType songType, std::string title);
 	void sortSongs(std::vector<int> numbers);
 	size_t getAllStepCount(int songNum, int loopCnt) const;
+	/*----- Bookmark -----*/
+	void addBookmark(int songNum, std::string name, int order, int step);
+	void changeBookmark(int songNum, int i, std::string name, int order, int step);
+	void removeBookmark(int songNum, int i);
+	void clearBookmark(int songNum);
+	void swapBookmarks(int songNum, int a, int b);
+	void sortBookmarkByPosition(int songNum);
+	void sortBookmarkByName(int songNum);
+	Bookmark getBookmark(int songNum, int i) const;
+	std::vector<int> findBookmarks(int songNum, int order, int step);
+	Bookmark getPreviousBookmark(int songNum, int order, int step);
+	Bookmark getNextBookmark(int songNum, int order, int step);
+	size_t getBookmarkSize(int songNum) const;
 	/*----- Track -----*/
 	void setEffectDisplayWidth(int songNum, int trackNum, size_t w);
 	size_t getEffectDisplayWidth(int songNum, int trackNum) const;