+++ BambooTracker/gui/pattern_editor/pattern_editor_panel.hpp
@@ -39,7 +39,7 @@ public:
 
 	void changeEditable();
 	int getFullColmunSize() const;
-	void updatePositionByStepUpdate(bool isFirstUpdate);
+	void updatePositionByStepUpdate(bool isFirstUpdate, bool forceJump = false);
 	int getScrollableCountByTrack() const;
 
 	void copySelectedCells();
@@ -110,6 +110,8 @@ signals:
 	void currentTrackChanged(int num);
 	void currentOrderChanged(int num, int max);
 	void effectColsCompanded(int num, int max);
+	void bookmarkToggleRequested(int order, int step);
+	void bookmarkJumpRequested(bool toNext, int order, int step);
 
 	void selected(bool isSelected);
 	void returnPressed();